requirements.txt
# 기본 API/서버/인프라 관련 (2025-07 기준 최신 안정 버전)
openai>=1.14.3
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.4
websockets>=12.0
aiofiles>=23.2.1

# 수학/과학 연산 라이브러리
numpy>=1.26.4
scipy>=1.13.0

# 오디오 처리
sounddevice>=0.4.7
soundfile>=0.12.1
webrtcvad @ git+https://github.com/wiseman/py-webrtcvad.git

# Whisper (OpenAI GitHub 최신 버전)
openai-whisper @ git+https://github.com/openai/whisper.git